requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
numpy==1.24.3
torch==2.0.1
scikit-learn==1.3.0
requests==2.31.0
python-dotenv==1.0.0
plotly==5.16.1
dash==2.11.1
dash-bootstrap-components==1.4.2
python-binance==1.0.19
newsapi-python==0.2.7 